+++ packages/react-ui/src/components/organisms/AppBar/BurgerMenu.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import { Divider, Hidden, IconButton } from '@mui/material';
-import { styled } from '@mui/material/styles';
+import { alpha, styled } from '@mui/material/styles';
 import { MenuOutlined } from '@mui/icons-material';
 
 import { APPBAR_SIZE } from '../../../theme/themeConstants';
@@ -16,18 +16,22 @@ const MenuButton = styled(IconButton)(({ theme }) => ({
   marginRight: theme.spacing(1),
 
   '&.MuiButtonBase-root svg path': {
-    fill: theme.palette.background.paper
+    fill: theme.palette.brand.appBarContrastText
   }
 }));
 
-export default function BurguerMenu({ onClickMenu }) {
+const MenuDivider = styled(Divider)(({ theme }) => ({
+  borderColor: alpha(theme.palette.brand.appBarContrastText, 0.12)
+}));
+
+export default function BurgerMenu({ onClickMenu }) {
   return (
     <Hidden mdUp>
       <Menu>
         <MenuButton onClick={onClickMenu}>
           <MenuOutlined />
         </MenuButton>
-        <Divider orientation='vertical' flexItem light />
+        <MenuDivider orientation='vertical' flexItem />
       </Menu>
     </Hidden>
   );